Finalize transition from models.markers
